+++ app/src/main/java/com/depromeet/team6/presentation/model/course/LegInfo.kt
@@ -4,10 +4,10 @@ import androidx.compose.ui.graphics.Color
 
 data class LegInfo(
     val transportType: TransportType,
-    val subTypeIdx : Int = 0,
-    val sectionTime : Int,
-    val distance : Int,
-    val routeColor : Color,
-    val startPoint : WayPoint,
+    val subTypeIdx: Int = 0,
+    val sectionTime: Int,
+    val distance: Int,
+    val routeColor: Color,
+    val startPoint: WayPoint,
     val endPoint: WayPoint
 )
