+++ app/src/main/java/com/depromeet/team6/presentation/ui/course/component/LastTransportInfoItem.kt
@@ -1,4 +1,4 @@
-package com.depromeet.team6.presentation.ui.course_search.component
+package com.depromeet.team6.presentation.ui.course.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -248,7 +248,7 @@ fun LastTransportInfoItemPreview() {
             ),
             routeColor = defaultTeam6Colors.black,
             distance = 10
-        ),
+        )
     )
 
     val mockData = LastTransportInfo(
@@ -331,7 +331,7 @@ fun LastTransportInfoItemPreview2() {
             ),
             routeColor = defaultTeam6Colors.black,
             distance = 10
-        ),
+        )
     )
 
     val mockData = LastTransportInfo(
