+++ app/src/main/java/com/depromeet/team6/presentation/ui/main/MainActivity.kt
@@ -14,7 +14,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.depromeet.team6.presentation.ui.course_search.CourseScreen
+import com.depromeet.team6.presentation.ui.course.CourseScreen
 import com.depromeet.team6.ui.theme.Team6Theme
 import com.depromeet.team6.ui.theme.Team6Theme.colors
 import com.depromeet.team6.ui.theme.Team6Theme.typography
